Feature/update

+++ DraughtsGameFunctionModels/Controller/GetPlayerTips.cs
@@ -6,8 +6,8 @@ namespace DraughtsGameFunctionModels.Controller
 {
     public class GetPlayerTips
     {
-        public int[,] Board { get; set; }
+        public Int64[,] Board { get; set; }
 
-        public int TipFor { get; set; }
+        public Int64 TipFor { get; set; }
     }
 }
