+++ DraughtsGameFunctionModels/Service/Piece.cs
@@ -6,8 +6,8 @@ namespace DraughtsGameFunctionModels.Service
 {
     public class Piece
     {
-        public int Height { get; set; }
+        public Int64 Height { get; set; }
 
-        public int Width { get; set; }
+        public Int64 Width { get; set; }
     }
 }
